pkg/types: Move Atom and SpecTuple from pkg/graph
