--- pkg/graph/spec.go
@@ -1,5 +1,0 @@
-package graph
-
-func (st SpecTuple) String() string {
-	return st.Host + ":" + st.Target
-}
